Auth, VPwrapper and HPwrapper
