+++ index.js
@@ -1,60 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
-import Icon from '@material-ui/core/Icon';
-import { AccessAlarm, ExpandMore, OpenInNew, Timeline, GetApp, ChevronRight, Add } from '@material-ui/icons';
-import Grid from '@material-ui/core/Grid';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
-import axios from 'axios';
-import PropTypes from 'prop-types';
-import Divider from '@material-ui/core/Divider';
-import LockIcon from '@material-ui/icons/Lock';
-import { makeStyles, withStyles } from '@material-ui/core/styles';
-import { borders } from '@material-ui/system';
 import '@fontsource/roboto';
 import {Helmet} from "react-helmet";
-import Select from '@material-ui/core/Select';
-import { createMuiTheme } from '@material-ui/core/styles';
 import { BrowserRouter as Router, Switch, Route, Redirect } from 'react-router-dom';
 import './App.css';
 
-import BQPane from './pages/BuildQuery/BuildQueryPane.js';
-
-//import OKMAP from './ViewPane.js';
-import useStyles from './useStyles.js';
-import ViewPane from './ViewPane.js';
-import AboutUs from './components/AboutUs';
 import MainPanel from './MainPanel.js';
 import TopNav from './TopNav.js';
-
-import { makeRequest } from './request/CancerDataManagement.js';
-//import ClientAddFilter from './ClientAddFilter.js'
-
-//These are just some CSS props
-const defaultProps = {
-  m: 1,
-};
-
-const boxProps = {
-  border: 3,
-};
-
-const boxProps_padding = {
-  border: 3,
-  margin: 1,
-  paddingRight: '10px',
-  display: 'inline-block',
-};
-
 //Versioning
 var version = "0.1";
-
-//Global Variables
-var flag = 0;
 
 var localurl = "/material-app";
 var serverurl = "/ICGS/Oncosplice/testing";
@@ -64,12 +18,6 @@ const targeturl = serverurl;
 
 var routeurl = "/ICGS/Oncosplice/testing/index.html"
 
-var queryhistory_dat = [];
-
-function none()
-{
-  return null;
-}
 //App function is the top level function, all components are underneath this function. At this level the routing is controlled.
 function App() {
